drop VersionedDependency
